02_movies_admin/requirements.txt
asgiref==3.11.1
Django==5.1
django-cors-headers==4.9.0
django-split-settings==1.1.0
Faker==13.15.0
flake8==4.0.0
mccabe==0.6.1
packaging==26.0
psycopg2-binary==2.9.11
pycodestyle==2.8.0
pyflakes==2.4.0
python-dateutil==2.9.0.post0
python-dotenv==0.20.0
pytz==2026.1.post1
setuptools==82.0.1
six==1.17.0
sqlparse==0.5.5
wheel==0.46.3
uwsgi==2.0.28
